feat: disable linux removal check
When uninstalling a kernel normally the prerm script checks you are not removing the currently running kernel vresion. This is a safeguard against removing all kernels on a system. Unfortutenaly when this check is triggered it runs in a console menu which is hidden for the user of this script, therefore we disable this check as a user already has to confirm twice